feat: Add ReadingOrderEvaluator for new reading-order model (#29)
* correct mpy


* reformatting


* adding the script to make an initial dataset from pdf's


* before switching to specific docling-core branch


* rebased on kv-items and updated the create script in CVAT


* fixed the cvat


* added the annotation description on CVAT


* added the annotation description on CVAT (2)


* added the annotation description on CVAT (3)


* feat: Add new reading-order model evaluator, re-factor to BaseReadingOrderEvaluator


* Rename New


* fix: Fix calling DoclingDocument methods. Pin latest versions of docling and docling-core.
Remove commented out code, remove unused imports, code formatting.


* docs: Re-run the evaluations for DPBench and update the docs


* docs: Re-run the evaluations for OmniDocBench and update the docs


* docs: Re-run the evaluations for DPBench and update the docs


* docs: Re-run the evaluations for OmniDocBench and update the docs


* docs: Minor improvements in DPBench OmniDocBench documentation


---------
