--- docs/evaluations/DPBench/evaluation_DPBench_layout_mAP[0.5_0.95].txt
@@ -1,43 +1,0 @@
-DPBench layout mAP[0.5_0.95]:
-
-mean=0.62 median=0.65 std=0.20
-
-| mAP[0.5_0.95]   |   prob [%] |   acc [%] |   1-acc [%] |   total |
-|-----------------|------------|-----------|-------------|---------|
-| (0.000, 0.050]  |        0.5 |       0   |       100   |       1 |
-| (0.050, 0.100]  |        0   |       0.5 |        99.5 |       0 |
-| (0.100, 0.150]  |        0   |       0.5 |        99.5 |       0 |
-| (0.150, 0.200]  |        1.5 |       0.5 |        99.5 |       3 |
-| (0.200, 0.250]  |        1   |       2   |        98   |       2 |
-| (0.250, 0.300]  |        2.5 |       3   |        97   |       5 |
-| (0.300, 0.350]  |        5.5 |       5.5 |        94.5 |      11 |
-| (0.350, 0.400]  |        4.5 |      11   |        89   |       9 |
-| (0.400, 0.450]  |        8.5 |      15.5 |        84.5 |      17 |
-| (0.450, 0.500]  |        4.5 |      24   |        76   |       9 |
-| (0.500, 0.550]  |        5.5 |      28.5 |        71.5 |      11 |
-| (0.550, 0.600]  |        6   |      34   |        66   |      12 |
-| (0.600, 0.650]  |        8.5 |      40   |        60   |      17 |
-| (0.650, 0.700]  |       10.5 |      48.5 |        51.5 |      21 |
-| (0.700, 0.750]  |       11   |      59   |        41   |      22 |
-| (0.750, 0.800]  |        9   |      70   |        30   |      18 |
-| (0.800, 0.850]  |       11.5 |      79   |        21   |      23 |
-| (0.850, 0.900]  |        5.5 |      90.5 |         9.5 |      11 |
-| (0.900, 0.950]  |        4   |      96   |         4   |       8 |
-| (0.950, 1.000]  |        0   |     100   |         0   |       0 |
-
-
-Class mAP[0.5:0.95] table:
-
-| label          |   Class mAP[0.5:0.95] |
-|----------------|-----------------------|
-| table          |                 89.08 |
-| picture        |                 76.1  |
-| document_index |                 75.52 |
-| text           |                 67.8  |
-| caption        |                 45.8  |
-| section_header |                 44.26 |
-| page_footer    |                 34.42 |
-| list_item      |                 29.04 |
-| footnote       |                 22.08 |
-| page_header    |                 15.11 |
-| formula        |                  6.62 |
