+++ docs/evaluations/DPBench/evaluation_DPBench_layout_mAP_0.5_0.95.txt
@@ -1,0 +1,45 @@
+DPBench layout mAP_0.5_0.95: mean=0.48 median=0.48 std=0.19
+
+| mAP_0.5_0.95   |   prob [%] |   acc [%] |   1-acc [%] |   total |
+|----------------|------------|-----------|-------------|---------|
+| (0.000, 0.050] |        1   |       0   |       100   |       2 |
+| (0.050, 0.100] |        0.5 |       1   |        99   |       1 |
+| (0.100, 0.150] |        0.5 |       1.5 |        98.5 |       1 |
+| (0.150, 0.200] |        3   |       2   |        98   |       6 |
+| (0.200, 0.250] |        5   |       5   |        95   |      10 |
+| (0.250, 0.300] |        6   |      10   |        90   |      12 |
+| (0.300, 0.350] |        9.5 |      16   |        84   |      19 |
+| (0.350, 0.400] |       10.5 |      25.5 |        74.5 |      21 |
+| (0.400, 0.450] |        8   |      36   |        64   |      16 |
+| (0.450, 0.500] |       11.5 |      44   |        56   |      23 |
+| (0.500, 0.550] |       11   |      55.5 |        44.5 |      22 |
+| (0.550, 0.600] |        8.5 |      66.5 |        33.5 |      17 |
+| (0.600, 0.650] |        5   |      75   |        25   |      10 |
+| (0.650, 0.700] |        6.5 |      80   |        20   |      13 |
+| (0.700, 0.750] |        4.5 |      86.5 |        13.5 |       9 |
+| (0.750, 0.800] |        1.5 |      91   |         9   |       3 |
+| (0.800, 0.850] |        5.5 |      92.5 |         7.5 |      11 |
+| (0.850, 0.900] |        1.5 |      98   |         2   |       3 |
+| (0.900, 0.950] |        0.5 |      99.5 |         0.5 |       1 |
+| (0.950, 1.000] |        0   |     100   |         0   |       0 |
+
+
+
+
+
+AP[0.5:0.05:0.95] per class (reported as %):
+
+| label          |   Class mAP[0.5:0.95] |
+|----------------|-----------------------|
+| table          |                 89.35 |
+| document_index |                 75.52 |
+| picture        |                 75.42 |
+| text           |                 60.95 |
+| section_header |                 39.91 |
+| list_item      |                 24.59 |
+| footnote       |                 18.63 |
+| caption        |                  9.61 |
+| formula        |                  6.49 |
+| page_footer    |                  1.49 |
+
+Total mAP[0.5:0.05:0.95] (reported as %): 40.20
