--- docs/evaluations/DPBench/evaluation_DPBench_markdown_text_BLEU.txt
@@ -1,26 +1,0 @@
-DPBench markdown_text BLEU:
-
-mean=0.81 median=0.87 std=0.20
-
-| BLEU           |   prob [%] |   acc [%] |   1-acc [%] |   total |
-|----------------|------------|-----------|-------------|---------|
-| (0.000, 0.050] |        1   |       0   |       100   |       2 |
-| (0.050, 0.100] |        0.5 |       1   |        99   |       1 |
-| (0.100, 0.150] |        0.5 |       1.5 |        98.5 |       1 |
-| (0.150, 0.200] |        1.5 |       2   |        98   |       3 |
-| (0.200, 0.250] |        1   |       3.5 |        96.5 |       2 |
-| (0.250, 0.300] |        0   |       4.5 |        95.5 |       0 |
-| (0.300, 0.350] |        0.5 |       4.5 |        95.5 |       1 |
-| (0.350, 0.400] |        0   |       5   |        95   |       0 |
-| (0.400, 0.450] |        0.5 |       5   |        95   |       1 |
-| (0.450, 0.500] |        0.5 |       5.5 |        94.5 |       1 |
-| (0.500, 0.550] |        3.5 |       6   |        94   |       7 |
-| (0.550, 0.600] |        1   |       9.5 |        90.5 |       2 |
-| (0.600, 0.650] |        4   |      10.5 |        89.5 |       8 |
-| (0.650, 0.700] |        2   |      14.5 |        85.5 |       4 |
-| (0.700, 0.750] |        3.5 |      16.5 |        83.5 |       7 |
-| (0.750, 0.800] |       10   |      20   |        80   |      20 |
-| (0.800, 0.850] |        9.5 |      30   |        70   |      19 |
-| (0.850, 0.900] |       21   |      39.5 |        60.5 |      42 |
-| (0.900, 0.950] |       22.5 |      60.5 |        39.5 |      45 |
-| (0.950, 1.000] |       17   |      83   |        17   |      34 |
