+++ docs/evaluations/DPBench/evaluation_DPBench_reading_order_ARD_norm.txt
@@ -1,6 +1,4 @@
-DPBench reading_order ARD_norm:
-
-mean=0.98 median=1.00 std=0.05
+DPBench reading_order ARD_norm: mean=0.99 median=1.00 std=0.02
 
 | ARD_norm       |   prob [%] |   acc [%] |   1-acc [%] |   total |
 |----------------|------------|-----------|-------------|---------|
@@ -17,10 +15,12 @@ mean=0.98 median=1.00 std=0.05
 | (0.500, 0.550] |        0   |       0   |       100   |       0 |
 | (0.550, 0.600] |        0   |       0   |       100   |       0 |
 | (0.600, 0.650] |        0   |       0   |       100   |       0 |
-| (0.650, 0.700] |        1   |       0   |       100   |       2 |
-| (0.700, 0.750] |        0.5 |       1   |        99   |       1 |
-| (0.750, 0.800] |        1   |       1.5 |        98.5 |       2 |
-| (0.800, 0.850] |        2.5 |       2.5 |        97.5 |       5 |
-| (0.850, 0.900] |        0.5 |       5   |        95   |       1 |
-| (0.900, 0.950] |        1.5 |       5.5 |        94.5 |       3 |
-| (0.950, 1.000] |       93   |       7   |        93   |     186 |
+| (0.650, 0.700] |        0   |       0   |       100   |       0 |
+| (0.700, 0.750] |        0   |       0   |       100   |       0 |
+| (0.750, 0.800] |        0   |       0   |       100   |       0 |
+| (0.800, 0.850] |        0.5 |       0   |       100   |       1 |
+| (0.850, 0.900] |        0   |       0.5 |        99.5 |       0 |
+| (0.900, 0.950] |        4   |       0.5 |        99.5 |       8 |
+| (0.950, 1.000] |       95.5 |       4.5 |        95.5 |     191 |
+
+
