--- docs/evaluations/OmniDocBench/evaluation_OmniDocBench_layout_mAP[0.5_0.95].txt
@@ -1,26 +1,0 @@
-OmniDocBench layout mAP[0.5_0.95]:
-
-mean=0.32 median=0.31 std=0.22
-
-| mAP[0.5_0.95]   |   prob [%] |   acc [%] |   1-acc [%] |   total |
-|-----------------|------------|-----------|-------------|---------|
-| (0.000, 0.050]  |      15.23 |      0    |      100    |     143 |
-| (0.050, 0.100]  |       4.05 |     15.23 |       84.77 |      38 |
-| (0.100, 0.150]  |       5.43 |     19.28 |       80.72 |      51 |
-| (0.150, 0.200]  |       7.45 |     24.71 |       75.29 |      70 |
-| (0.200, 0.250]  |       7.03 |     32.16 |       67.84 |      66 |
-| (0.250, 0.300]  |       7.35 |     39.19 |       60.81 |      69 |
-| (0.300, 0.350]  |      10.01 |     46.54 |       53.46 |      94 |
-| (0.350, 0.400]  |       6.92 |     56.55 |       43.45 |      65 |
-| (0.400, 0.450]  |       8.95 |     63.47 |       36.53 |      84 |
-| (0.450, 0.500]  |       7.03 |     72.42 |       27.58 |      66 |
-| (0.500, 0.550]  |       5.96 |     79.45 |       20.55 |      56 |
-| (0.550, 0.600]  |       5.11 |     85.41 |       14.59 |      48 |
-| (0.600, 0.650]  |       3.73 |     90.52 |        9.48 |      35 |
-| (0.650, 0.700]  |       1.81 |     94.25 |        5.75 |      17 |
-| (0.700, 0.750]  |       1.28 |     96.06 |        3.94 |      12 |
-| (0.750, 0.800]  |       0.85 |     97.34 |        2.66 |       8 |
-| (0.800, 0.850]  |       0.53 |     98.19 |        1.81 |       5 |
-| (0.850, 0.900]  |       0.53 |     98.72 |        1.28 |       5 |
-| (0.900, 0.950]  |       0.32 |     99.25 |        0.75 |       3 |
-| (0.950, 1.000]  |       0.43 |     99.57 |        0.43 |       4 |
