+++ docs/evaluations/OmniDocBench/evaluation_OmniDocBench_layout_mAP_0.5_0.95.txt
@@ -1,0 +1,44 @@
+OmniDocBench layout mAP_0.5_0.95: mean=0.24 median=0.23 std=0.18
+
+| mAP_0.5_0.95   |   prob [%] |   acc [%] |   1-acc [%] |   total |
+|----------------|------------|-----------|-------------|---------|
+| (0.000, 0.050] |      18.59 |      0    |      100    |     174 |
+| (0.050, 0.100] |       5.45 |     18.59 |       81.41 |      51 |
+| (0.100, 0.150] |       9.19 |     24.04 |       75.96 |      86 |
+| (0.150, 0.200] |       9.29 |     33.23 |       66.77 |      87 |
+| (0.200, 0.250] |      12.18 |     42.52 |       57.48 |     114 |
+| (0.250, 0.300] |      11.54 |     54.7  |       45.3  |     108 |
+| (0.300, 0.350] |       8.76 |     66.24 |       33.76 |      82 |
+| (0.350, 0.400] |       6.94 |     75    |       25    |      65 |
+| (0.400, 0.450] |       6.52 |     81.94 |       18.06 |      61 |
+| (0.450, 0.500] |       4.59 |     88.46 |       11.54 |      43 |
+| (0.500, 0.550] |       3.42 |     93.06 |        6.94 |      32 |
+| (0.550, 0.600] |       1.28 |     96.47 |        3.53 |      12 |
+| (0.600, 0.650] |       0.43 |     97.76 |        2.24 |       4 |
+| (0.650, 0.700] |       0.32 |     98.18 |        1.82 |       3 |
+| (0.700, 0.750] |       0    |     98.5  |        1.5  |       0 |
+| (0.750, 0.800] |       0.21 |     98.5  |        1.5  |       2 |
+| (0.800, 0.850] |       0    |     98.72 |        1.28 |       0 |
+| (0.850, 0.900] |       0.64 |     98.72 |        1.28 |       6 |
+| (0.900, 0.950] |       0.21 |     99.36 |        0.64 |       2 |
+| (0.950, 1.000] |       0.43 |     99.57 |        0.43 |       4 |
+
+
+
+
+
+AP[0.5:0.05:0.95] per class (reported as %):
+
+| label          |   Class mAP[0.5:0.95] |
+|----------------|-----------------------|
+| table          |                 69.31 |
+| picture        |                 29.45 |
+| text           |                 20.19 |
+| section_header |                 12.42 |
+| caption        |                  4.22 |
+| formula        |                  3.91 |
+| footnote       |                  2.46 |
+| page_footer    |                  0.45 |
+| page_header    |                  0    |
+
+Total mAP[0.5:0.05:0.95] (reported as %): 15.82
