--- docs/evaluations/OmniDocBench/evaluation_OmniDocBench_markdown_text_BLEU.txt
@@ -1,26 +1,0 @@
-OmniDocBench markdown_text BLEU:
-
-mean=0.30 median=0.11 std=0.33
-
-| BLEU           |   prob [%] |   acc [%] |   1-acc [%] |   total |
-|----------------|------------|-----------|-------------|---------|
-| (0.000, 0.050] |      41.59 |      0    |      100    |     408 |
-| (0.050, 0.100] |       6.83 |     41.59 |       58.41 |      67 |
-| (0.100, 0.150] |       4.18 |     48.42 |       51.58 |      41 |
-| (0.150, 0.200] |       3.26 |     52.6  |       47.4  |      32 |
-| (0.200, 0.250] |       2.45 |     55.86 |       44.14 |      24 |
-| (0.250, 0.300] |       1.83 |     58.31 |       41.69 |      18 |
-| (0.300, 0.350] |       1.83 |     60.14 |       39.86 |      18 |
-| (0.350, 0.400] |       2.04 |     61.98 |       38.02 |      20 |
-| (0.400, 0.450] |       2.04 |     64.02 |       35.98 |      20 |
-| (0.450, 0.500] |       2.55 |     66.06 |       33.94 |      25 |
-| (0.500, 0.550] |       2.04 |     68.6  |       31.4  |      20 |
-| (0.550, 0.600] |       2.04 |     70.64 |       29.36 |      20 |
-| (0.600, 0.650] |       2.75 |     72.68 |       27.32 |      27 |
-| (0.650, 0.700] |       2.96 |     75.43 |       24.57 |      29 |
-| (0.700, 0.750] |       4.69 |     78.39 |       21.61 |      46 |
-| (0.750, 0.800] |       4.28 |     83.08 |       16.92 |      42 |
-| (0.800, 0.850] |       4.79 |     87.36 |       12.64 |      47 |
-| (0.850, 0.900] |       4.59 |     92.15 |        7.85 |      45 |
-| (0.900, 0.950] |       2.65 |     96.74 |        3.26 |      26 |
-| (0.950, 1.000] |       0.61 |     99.39 |        0.61 |       6 |
