+++ docs/evaluations/OmniDocBench/evaluation_OmniDocBench_reading_order_ARD_norm.txt
@@ -1,6 +1,4 @@
-OmniDocBench reading_order ARD_norm:
-
-mean=0.84 median=0.84 std=0.12
+OmniDocBench reading_order ARD_norm: mean=0.85 median=0.85 std=0.11
 
 | ARD_norm       |   prob [%] |   acc [%] |   1-acc [%] |   total |
 |----------------|------------|-----------|-------------|---------|
@@ -14,13 +12,15 @@ mean=0.84 median=0.84 std=0.12
 | (0.350, 0.400] |       0    |      0    |      100    |       0 |
 | (0.400, 0.450] |       0    |      0    |      100    |       0 |
 | (0.450, 0.500] |       0    |      0    |      100    |       0 |
-| (0.500, 0.550] |       1.53 |      0    |      100    |      15 |
-| (0.550, 0.600] |       2.24 |      1.53 |       98.47 |      22 |
-| (0.600, 0.650] |       2.55 |      3.77 |       96.23 |      25 |
-| (0.650, 0.700] |       4.89 |      6.32 |       93.68 |      48 |
-| (0.700, 0.750] |       8.15 |     11.21 |       88.79 |      80 |
-| (0.750, 0.800] |      17.74 |     19.37 |       80.63 |     174 |
-| (0.800, 0.850] |      17.43 |     37.1  |       62.9  |     171 |
-| (0.850, 0.900] |      17.13 |     54.54 |       45.46 |     168 |
-| (0.900, 0.950] |       7.44 |     71.66 |       28.34 |      73 |
-| (0.950, 1.000] |      20.9  |     79.1  |       20.9  |     205 |
+| (0.500, 0.550] |       0.92 |      0    |      100    |       9 |
+| (0.550, 0.600] |       0.82 |      0.92 |       99.08 |       8 |
+| (0.600, 0.650] |       1.22 |      1.73 |       98.27 |      12 |
+| (0.650, 0.700] |       4.28 |      2.96 |       97.04 |      42 |
+| (0.700, 0.750] |       7.75 |      7.24 |       92.76 |      76 |
+| (0.750, 0.800] |      15.9  |     14.98 |       85.02 |     156 |
+| (0.800, 0.850] |      20.39 |     30.89 |       69.11 |     200 |
+| (0.850, 0.900] |      19.57 |     51.27 |       48.73 |     192 |
+| (0.900, 0.950] |       7.95 |     70.85 |       29.15 |      78 |
+| (0.950, 1.000] |      21.2  |     78.8  |       21.2  |     208 |
+
+
