+++ docs/evaluations/OmniDocBench/evaluation_OmniDocBench_reading_order_weighted_ARD.txt
@@ -1,6 +1,4 @@
-OmniDocBench reading_order weighted_ARD:
-
-mean=0.98 median=0.99 std=0.03
+OmniDocBench reading_order weighted_ARD: mean=0.98 median=0.99 std=0.03
 
 | weighted_ARD   |   prob [%] |   acc [%] |   1-acc [%] |   total |
 |----------------|------------|-----------|-------------|---------|
@@ -21,6 +19,8 @@ mean=0.98 median=0.99 std=0.03
 | (0.700, 0.750] |       0    |      0    |      100    |       0 |
 | (0.750, 0.800] |       0.41 |      0    |      100    |       4 |
 | (0.800, 0.850] |       0.1  |      0.41 |       99.59 |       1 |
-| (0.850, 0.900] |       2.04 |      0.51 |       99.49 |      20 |
-| (0.900, 0.950] |       6.22 |      2.55 |       97.45 |      61 |
-| (0.950, 1.000] |      91.23 |      8.77 |       91.23 |     895 |
+| (0.850, 0.900] |       1.83 |      0.51 |       99.49 |      18 |
+| (0.900, 0.950] |       6.73 |      2.34 |       97.66 |      66 |
+| (0.950, 1.000] |      90.93 |      9.07 |       90.93 |     892 |
+
+
