+++ pyproject.toml
@@ -31,7 +31,7 @@ lxml = "^5.3.0"
 datasets = "^3.2.0"
 apted = "^1.0.3"
 Distance = "^0.1.3"
-docling = "^2.21.0"
+docling = "^2.24.0"
 matplotlib = "^3.10.0"
 torch = "^2.5.1"
 torchmetrics = "^1.6.0"
